# frontend/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Authentication","description":"Authentication APIs"},{"name":"Voting","description":"Vote APIs"},{"name":"Ranking","description":"Ranking feed APIs"},{"name":"Posts","description":"Post submission APIs"},{"name":"Moderator","description":"Moderator tools"},{"name":"Admin","description":"Admin tools"},{"name":"Notifications","description":"Notification APIs"},{"name":"Tags","description":"Tag APIs"},{"name":"Comments","description":"Comment APIs"}],"paths":{"/api/v1/admin/users/{id}/role":{"put":{"tags":["Admin"],"summary":"Change user role (admin only)","operationId":"changeUserRole","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"role","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserListItem"}}}}}}},"/api/v1/posts":{"get":{"tags":["Posts"],"summary":"List posts","operationId":"listPosts","parameters":[{"name":"sort","in":"query","required":false,"schema":{"type":"string","default":"new"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponsePagePostResponse"}}}}}},"post":{"tags":["Posts"],"summary":"Create a new post","operationId":"createPost","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePostRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponsePostDetailResponse"}}}}}}},"/api/v1/posts/{postId}/upvote":{"post":{"tags":["Voting"],"summary":"Upvote a post","operationId":"upvote","parameters":[{"name":"postId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseBoolean"}}}}}},"delete":{"tags":["Voting"],"summary":"Remove upvote from a post","operationId":"removeVote","parameters":[{"name":"postId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseBoolean"}}}}}}},"/api/v1/posts/{postId}/comments":{"get":{"tags":["Comments"],"summary":"Get comment tree for a post","operationId":"getComments","parameters":[{"name":"postId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListCommentResponse"}}}}}},"post":{"tags":["Comments"],"summary":"Create a comment on a post","operationId":"createComment","parameters":[{"name":"postId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCommentRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseCommentResponse"}}}}}}},"/api/v1/auth/register":{"post":{"tags":["Authentication"],"summary":"Register a new user","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthResponse"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["Authentication"],"summary":"Login and get JWT token","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthResponse"}}}}}}},"/api/v1/notifications/{id}/read":{"patch":{"tags":["Notifications"],"summary":"Mark notification as read","operationId":"markAsRead","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseBoolean"}}}}}}},"/api/v1/tags":{"get":{"tags":["Tags"],"summary":"List all tags","operationId":"listTags","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListString"}}}}}}},"/api/v1/tags/{tagName}/posts":{"get":{"tags":["Tags"],"summary":"Get posts by tag","operationId":"getPostsByTag","parameters":[{"name":"tagName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListPostResponse"}}}}}}},"/api/v1/ranking/top":{"get":{"tags":["Ranking"],"summary":"Get top posts by score","operationId":"getTop","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListPostResponse"}}}}}}},"/api/v1/ranking/new":{"get":{"tags":["Ranking"],"summary":"Get newest posts","operationId":"getNew","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListPostResponse"}}}}}}},"/api/v1/ranking/hot":{"get":{"tags":["Ranking"],"summary":"Get hot posts (HN style ranking)","operationId":"getHot","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListPostResponse"}}}}}}},"/api/v1/posts/{postId}/vote-status":{"get":{"tags":["Voting"],"summary":"Check if current user has voted","operationId":"voteStatus","parameters":[{"name":"postId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseBoolean"}}}}}}},"/api/v1/posts/{postId}/vote-count":{"get":{"tags":["Voting"],"summary":"Get vote count for a post","operationId":"voteCount","parameters":[{"name":"postId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseInteger"}}}}}}},"/api/v1/posts/{id}":{"get":{"tags":["Posts"],"summary":"Get post by ID","operationId":"getPost","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponsePostDetailResponse"}}}}}},"delete":{"tags":["Posts"],"summary":"Delete own post","operationId":"deletePost","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseBoolean"}}}}}}},"/api/v1/posts/user/{username}":{"get":{"tags":["Posts"],"summary":"Get posts by username","operationId":"getPostsByUser","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponsePagePostResponse"}}}}}}},"/api/v1/notifications":{"get":{"tags":["Notifications"],"summary":"Get user notifications","operationId":"getNotifications","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponsePageNotificationResponse"}}}}}}},"/api/v1/notifications/unread-count":{"get":{"tags":["Notifications"],"summary":"Get unread notification count","operationId":"getUnreadCount","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseLong"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["Authentication"],"summary":"Get current user info","operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserInfoResponse"}}}}}}},"/api/v1/admin/users":{"get":{"tags":["Admin"],"summary":"List all users (admin only)","operationId":"listUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListUserListItem"}}}}}}},"/api/v1/moderator/posts/{id}":{"delete":{"tags":["Moderator"],"summary":"Delete any post (moderator/admin)","operationId":"deletePost_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseBoolean"}}}}}}},"/api/v1/moderator/comments/{id}":{"delete":{"tags":["Moderator"],"summary":"Delete any comment (moderator/admin)","operationId":"deleteComment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseBoolean"}}}}}}},"/api/v1/comments/{commentId}":{"delete":{"tags":["Comments"],"summary":"Delete own comment","operationId":"deleteComment_1","parameters":[{"name":"commentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseBoolean"}}}}}}}},"components":{"schemas":{"ApiResponseUserListItem":{"required":["code","message"],"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/UserListItem"}}},"UserListItem":{"required":["email","id","karma","role","username"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"email":{"type":"string"},"karma":{"type":"integer","format":"int32"},"role":{"type":"string"}}},"CreatePostRequest":{"required":["title","type"],"type":"object","properties":{"title":{"maxLength":300,"minLength":0,"type":"string"},"type":{"type":"string","enum":["LINK","ASK"]},"url":{"type":"string"},"content":{"type":"string"}}},"ApiResponsePostDetailResponse":{"required":["code","message"],"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/PostDetailResponse"}}},"PostDetailResponse":{"required":["authorId","authorUsername","commentCount","createdAt","hotScore","id","score","tags","title","type"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"url":{"type":"string"},"content":{"type":"string"},"summary":{"type":"string"},"score":{"type":"integer","format":"int32"},"hotScore":{"type":"number","format":"double"},"commentCount":{"type":"integer","format":"int32"},"type":{"type":"string"},"authorId":{"type":"integer","format":"int64"},"authorUsername":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"tags":{"type":"array","items":{"type":"string"}}}},"ApiResponseBoolean":{"required":["code","message"],"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"boolean"}}},"CreateCommentRequest":{"required":["content"],"type":"object","properties":{"content":{"type":"string"},"parentId":{"type":"integer","format":"int64"}}},"ApiResponseCommentResponse":{"required":["code","message"],"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/CommentResponse"}}},"CommentResponse":{"required":["authorId","authorUsername","children","content","createdAt","deleted","depth","id","postId","score"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"postId":{"type":"integer","format":"int64"},"authorId":{"type":"integer","format":"int64"},"authorUsername":{"type":"string"},"parentId":{"type":"integer","format":"int64"},"content":{"type":"string"},"score":{"type":"integer","format":"int32"},"depth":{"type":"integer","format":"int32"},"deleted":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"},"children":{"type":"array","items":{"$ref":"#/components/schemas/CommentResponse"}}}},"RegisterRequest":{"required":["email","password","username"],"type":"object","properties":{"username":{"maxLength":50,"minLength":3,"type":"string"},"email":{"type":"string"},"password":{"maxLength":100,"minLength":6,"type":"string"}}},"ApiResponseAuthResponse":{"required":["code","message"],"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/AuthResponse"}}},"AuthResponse":{"required":["role","token","tokenType","username"],"type":"object","properties":{"token":{"type":"string"},"tokenType":{"type":"string"},"username":{"type":"string"},"role":{"type":"string"}}},"LoginRequest":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"ApiResponseListString":{"required":["code","message"],"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"array","items":{"type":"string"}}}},"ApiResponseListPostResponse":{"required":["code","message"],"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/PostResponse"}}}},"PostResponse":{"required":["authorUsername","commentCount","createdAt","hotScore","id","score","tags","title","type"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"url":{"type":"string"},"content":{"type":"string"},"summary":{"type":"string"},"score":{"type":"integer","format":"int32"},"hotScore":{"type":"number","format":"double"},"commentCount":{"type":"integer","format":"int32"},"type":{"type":"string"},"authorUsername":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"tags":{"type":"array","items":{"type":"string"}}}},"ApiResponsePagePostResponse":{"required":["code","message"],"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/PagePostResponse"}}},"PagePostResponse":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/PostResponse"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"offset":{"type":"integer","format":"int64"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"paged":{"type":"boolean"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"direction":{"type":"string"},"nullHandling":{"type":"string"},"ascending":{"type":"boolean"},"property":{"type":"string"},"ignoreCase":{"type":"boolean"}}},"ApiResponseInteger":{"required":["code","message"],"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"integer","format":"int32"}}},"ApiResponseListCommentResponse":{"required":["code","message"],"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/CommentResponse"}}}},"ApiResponsePageNotificationResponse":{"required":["code","message"],"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/PageNotificationResponse"}}},"NotificationResponse":{"required":["content","createdAt","id","isRead","type"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"type":{"type":"string","enum":["COMMENT_REPLY","POST_MILESTONE","SYSTEM"]},"content":{"type":"string"},"isRead":{"type":"boolean"},"relatedPostId":{"type":"integer","format":"int64"},"relatedCommentId":{"type":"integer","format":"int64"},"createdAt":{"type":"string","format":"date-time"}}},"PageNotificationResponse":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/NotificationResponse"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"ApiResponseLong":{"required":["code","message"],"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"integer","format":"int64"}}},"ApiResponseUserInfoResponse":{"required":["code","message"],"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/UserInfoResponse"}}},"UserInfoResponse":{"required":["email","id","karma","role","username"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"email":{"type":"string"},"karma":{"type":"integer","format":"int32"},"role":{"type":"string"}}},"ApiResponseListUserListItem":{"required":["code","message"],"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/UserListItem"}}}}}}}